fix more bitness bugs

+++ lib/include/rpmsg_env.h
@@ -183,7 +183,7 @@ int32_t env_strncmp(char *dest, const char *src, uint32_t len);
 #if defined(RL_USE_ENVIRONMENT_CONTEXT) && (RL_USE_ENVIRONMENT_CONTEXT == 1)
 uint32_t env_map_vatopa(void *env, void *address);
 #else
-uint32_t env_map_vatopa(void *address);
+uintptr_t env_map_vatopa(void *address);
 #endif
 
 /*!
@@ -200,7 +200,7 @@ uint32_t env_map_vatopa(void *address);
 #if defined(RL_USE_ENVIRONMENT_CONTEXT) && (RL_USE_ENVIRONMENT_CONTEXT == 1)
 void *env_map_patova(void *env, uint32_t address);
 #else
-void *env_map_patova(uint32_t address);
+void *env_map_patova(uintptr_t address);
 #endif
 
 /*!
