+++ lib/rpmsg_lite/porting/environment/linux_sim/rpmsg_env_linux_sim.cpp
@@ -285,7 +285,7 @@ void env_wmb(void)
  *
  * @param address
  */
-uint32_t env_map_vatopa(void *address)
+uintptr_t env_map_vatopa(void *address)
 {
 	return platform_vatopa(address);
 }
@@ -295,7 +295,7 @@ uint32_t env_map_vatopa(void *address)
  *
  * @param address
  */
-void *env_map_patova(uint32_t address)
+void *env_map_patova(uintptr_t address)
 {
 	return platform_patova(address);
 }
