+++ lib/rpmsg_lite/porting/platform/simulation_linux/rpmsg_platform.c
@@ -233,9 +233,9 @@ void platform_cache_disable(void)
  * Dummy implementation
  *
  */
-uint32_t platform_vatopa(void *addr)
+uintptr_t platform_vatopa(void *addr)
 {
-    return ((uint32_t)(char *)addr);
+    return ((uintptr_t)(char *)addr);
 }
 
 /**
@@ -244,9 +244,9 @@ uint32_t platform_vatopa(void *addr)
  * Dummy implementation
  *
  */
-void *platform_patova(uint32_t addr)
+void *platform_patova(uintptr_t addr)
 {
-    return ((void *)(uint32_t)addr);
+    return ((void *)(uintptr_t)addr);
 }
 
 /**
